feat: Implement high-level POSIX message queue API
Implements a high-level API on top of POSIX message queues
(mq_overview(7)).

This API can be used to perform local RPC between processes that need
to exchange messages *fast* (or *easy*) with priority ordering.

The methods are mostly documented but there are still two corner cases
that need to be looked at and a lot of tests missing.